Index redirects to docs entry
home splash screen refactored into a separate file so it isn't lost forever
remove some dead wood
